Fixed warnings and added comments.

+++ win32/ffmpeg/libavutil/common.h
@@ -122,7 +122,7 @@ static av_always_inline av_const int av_clip_c(int a, int amin, int amax)
 static av_always_inline av_const uint8_t av_clip_uint8_c(int a)
 {
     if (a&(~0xFF)) return (-a)>>31;
-    else           return a;
+    else           return (uint8_t)a;
 }
 
 /**
@@ -133,7 +133,7 @@ static av_always_inline av_const uint8_t av_clip_uint8_c(int a)
 static av_always_inline av_const int8_t av_clip_int8_c(int a)
 {
     if ((a+0x80) & ~0xFF) return (a>>31) ^ 0x7F;
-    else                  return a;
+    else                  return (int8_t)a;
 }
 
 /**
@@ -144,7 +144,7 @@ static av_always_inline av_const int8_t av_clip_int8_c(int a)
 static av_always_inline av_const uint16_t av_clip_uint16_c(int a)
 {
     if (a&(~0xFFFF)) return (-a)>>31;
-    else             return a;
+    else             return (uint16_t)a;
 }
 
 /**
@@ -155,7 +155,7 @@ static av_always_inline av_const uint16_t av_clip_uint16_c(int a)
 static av_always_inline av_const int16_t av_clip_int16_c(int a)
 {
     if ((a+0x8000) & ~0xFFFF) return (a>>31) ^ 0x7FFF;
-    else                      return a;
+    else                      return (int16_t)a;
 }
 
 /**
@@ -166,7 +166,7 @@ static av_always_inline av_const int16_t av_clip_int16_c(int a)
 static av_always_inline av_const int32_t av_clipl_int32_c(int64_t a)
 {
     if ((a+0x80000000u) & ~UINT64_C(0xFFFFFFFF)) return (a>>63) ^ 0x7FFFFFFF;
-    else                                         return a;
+    else                                         return (int32_t)a;
 }
 
 /**
